# platform/api/openapi.yaml
openapi: 3.0.3
info:
  title: Platform PaaS API
  version: 0.3.14
  description: |
    Internal Platform API including AI-first MCP agent tokens and guarded command execution.
servers:
  - url: /api
    description: Platform API base
...
paths:
...
  /settings/integrations:
    get:
      tags: [Settings]
      summary: GitHub/GitLab OAuth apps and linking tokens
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Masked integration settings
    put:
      tags: [Settings]
      summary: Save GitHub/GitLab login and linking credentials
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Saved
  /settings/rotate-secrets:
    post:
      tags: [Settings]
      summary: One-click rotation of admin, database, and related secrets
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [confirm]
              properties:
                confirm:
                  type: string
                  example: ROTATE
      responses:
        '200':
          description: New values (shown once)
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Human JWT from POST /auth/login
    agentTokenAuth:
      type: http
      scheme: bearer
      description: Agent token (plat_agent_*)